Changed context to call and save coinData to use in other pages to save api calls
