+++ context/context.js
@@ -1,10 +1,16 @@
-import { createContext, useState, useEffect } from "react";
+import { createContext, useState, useEffect, useCallback } from "react";
 
 export const CoinMarketContext = createContext()
 
 export const CoinMarketProvider = ({children}) => {
+    const [top100Coins, setCoinData] = useState('')
+    const [coinMetaData, setCoinMetaData] = useState('')
 
-    const getTopTenCoins = async () => {
+    useEffect(() => {
+        setData();
+      }, []);
+
+      const getTopTenCoins = async () => {
         try {
             const res = await fetch('/api/getTopTen')
             const data = await res.json();
@@ -28,12 +34,39 @@ export const CoinMarketProvider = ({children}) => {
             console.log(e.message)
         }
     }
+    
+      const setData = useCallback(async () => {
+        let ids=[]
+        try {
+          let apiResponse = await getTopTenCoins();
+          let filteredResponse = [];
+    
+          for (let i = 0; i < apiResponse.length; i++) {
+            const element = apiResponse[i];
+              ids.push(element.id)
+              filteredResponse.push(element);
+          }
+          setCoinData(filteredResponse);
+        } catch (e) {
+          console.log(e.message);
+        }
+    
+        try {
+          let apiResponse = await getCoinMetaData(ids);
+          let filtered = Object.values(apiResponse).filter(item => item.id);
+          setCoinMetaData(filtered);
+        } catch (e) {
+          console.log(e.message)
+        }
+    
+      }, [getTopTenCoins, getCoinMetaData]);
+    
 
     return (
         <CoinMarketContext.Provider
         value = {{
-            getTopTenCoins,
-            getCoinMetaData
+            top100Coins,
+            coinMetaData
         }}
         >
             {children}
